checkplate {drfit}	R Documentation

Check raw data from a specified microtiter plate

Description

Report metadata from a specified microtiter plate from a specified database, box plot positive and negative (blind) controls, and show the response data on the plate.

Usage

  checkplate(plate,db="cytotox")

Arguments

plate 	 The number of the plate identifying it within the database.

db 	 The database to be used. Currently, the databases "cytotox" and "enzymes" of the UFT Department of Bioorganic Chemistry are supported (default is "cytotox").

Value

The function lists a report and shows two graphs.

Author(s)

Johannes Ranke jranke@uni-bremen.de http://www.uft.uni-bremen.de/chemie/ranke

Examples

# Check plate number 1 in the cytotox database
## Not run: data <- checkplate(1)

[Package drfit version 0.04-34 Index]