checksubstance {drfit}	R Documentation

Check raw data for a specified substance

Description

Report metadata for a specified substance from a specified database, and plot the data.

Usage

  checksubstance(substance,db="cytotox",experimentator="%",celltype="%",enzymetype="%",whereClause="1",ok="%")

Arguments

substance 	 The name of the substance identifying it within the database.

db 	 The database to be used. Currently, the databases "cytotox" and "enzymes" of the UFT Department of Bioorganic Chemistry are supported (default is "cytotox").

experimentator 	 The name of the experimentator whose data is to be used. Default is "%", which means that data from all experimentators are shown.

celltype 	 Currently, only data for IPC-81, C6, NB4, HeLa, Jurkat and U937 are supported. Default is "%", i.e. data for any cell type will be displayed.

enzymetype 	 Currently, only data for AChE, GR and GST are supported. The default value is "%", i.e. data for any enzyme type will be displayed.

whereClause 	 With this argument, additional conditions for the SQL query can be set, e.g. "where plate != 710". The default is 1 (in SQL syntax this means TRUE).

ok 	 With the default value "%", all data in the database is retrieved for the specified substance.

Value

The function lists a report and shows one graph.

Author(s)

Johannes Ranke jranke@uni-bremen.de http://www.uft.uni-bremen.de/chemie/ranke

Examples

# Check substance IM14 BF4 in the cytotox database
## Not run: checksubstance("IM14 BF4")

[Package drfit version 0.04-34 Index]