drdata {drfit}	R Documentation

Get dose-response data

Description

Get dose-response data from a remote mysql server

Usage

  drdata(substances, experimentator = "%", db = "cytotox", celltype = "IPC-81", 
    enzymetype="AChE", whereClause = "1", ok = "'ok','no fit'")

Arguments

substances 	 A string or an array of strings with the substance names for which dose-response data is to be retrieved.

experimentator 	 The name of the experimentator whose data is to be used. Default is " means that data from all experimentators are retrieved.

db 	 The database to be used. Currently, the databases "cytotox" and "enzymes" of the UFT Department of Bioorganic Chemistry are supported (default is "cytotox").

celltype 	 Currently, only data for IPC-81, C6, NB4, HeLa, Jurkat and U937 are supported.

enzymetype 	 Currently, only data for AChE, GR and GST are supported.

whereClause 	 With this argument, additional conditions for the SQL query can be set, e.g. "where plate != 710". The default is 1 (in SQL syntax this means TRUE).

ok 	 With the default value "'ok','no fit'", only data that has been checked and set to "ok" or "no fit" in the database is retrieved. The argument "no fit" will result in not using the data for fitting, but it will be plotted. Another sensible argument would be "'ok','no fit','?'", in order to additionally retrieve data which has not yet been checked.

Details

The function is currently only used for retrieving data from the mysql database "cytotox" of the UFT Department of Bioorganic Chemistry. Access to this database is limited to UFT staff. Additionally to the installation of the RODBC package, it is required to set up a ODBC data source with the name "cytotox", using an ODBC driver for mysql, probably myODBC. Then, under Unix, you can use iodbc or unixodbc for setting up the respective data source with data source name (DSN) "cytotox". For my setting using unixodbc, I am using the file ‘/etc/odbcinst.ini’ containing:

[MySQL]

Description 	 = 	 MySQL driver for ODBC

Driver 	 = 	 /usr/local/lib/libmyodbc.so

Setup 	 = 	 /usr/lib/odbc/libodbcmyS.so

and the file ‘/etc/odbc.ini’ containing:

[cytotox]

Description 	 = 	 Cytotoxicity database of the department of bioorganic chemistry, UFT Bremen

Driver 	 = 	 MySQL

Trace 	 = 	 Yes

TraceFile 	 = 	 /tmp/odbc.log

Database 	 = 	 cytotox

Server 	 = 	 eckehaat

Port 	 = 	 3306

.

Value

data 	 A data frame with a factor describing the dose levels, the numeric dose levels and a numeric column describing the response, as well as the entries for plate, experimentator, performed (date of test performance), celltype, unit (of the dose/concentration), and for the ok field in the database.

Author(s)

Johannes Ranke jranke@uni-bremen.de http://www.uft.uni-bremen.de/chemie/ranke

Examples

# Get cytotoxicity data for Tributyltin and zinc pyrithione, tested with IPC-81 cells
## Not run: data <- drdata(c("TBT","ZnPT2"))

[Package drfit version 0.04-34 Index]