linlogitf {drfit}	R Documentation

Linear-logistic function

Description

Helper function describing a special type of dose-response curves, showing a stimulus at subtoxic doses.

Usage

  linlogitf(x,k,f,mu,b)

Arguments

x 	 In this context, the x variable is the dose.

k 	 In the drfit functions, k is set to 1.

f 	 One of the parameters describing the curve shape.

mu 	 The parameter describing the location of the curve (log EC50).

b 	 One of the parameters describing the curve shape.

Value

The response at dose x.

Author(s)

Johannes Ranke jranke@uni-bremen.de http://www.uft.uni-bremen.de/chemie/ranke

References

van Ewijk, P. H. and Hoekstra, J. A. (1993) Ecotox Environ Safety 25 25-32

[Package drfit version 0.04-34 Index]